+++ test/src/com/newrelic/metrics/publish/processors/EpochProcessorTest.java
@@ -8,10 +8,10 @@ public class EpochProcessorTest {
 
     @Test
     public void testWithTwoValidPollCycles() throws InterruptedException {
-        EpochProcessor counter = new EpochProcessor();
-        Number firstProcess = counter.process(5);
-        resetTimer(counter, 1);
-        Number secondProcess = counter.process(6);
+        EpochProcessor processor = new EpochProcessor();
+        Number firstProcess = processor.process(5);
+        resetTimer(processor, 1);
+        Number secondProcess = processor.process(6);
 
         assertNull(firstProcess);
         assertEquals(1.0, secondProcess.doubleValue(), 0.1);
@@ -19,14 +19,14 @@ public class EpochProcessorTest {
 
     @Test
     public void testWithNullSecondPollCycle() throws InterruptedException {
-        EpochProcessor counter = new EpochProcessor();
-        Number firstProcess = counter.process(5);
-        resetTimer(counter, 1);
-        Number secondProcess = counter.process(null);
-        resetTimer(counter, 1);
-        Number thirdProcess = counter.process(6);
-        resetTimer(counter, 1);
-        Number fourthProcess = counter.process(7);
+        EpochProcessor processor = new EpochProcessor();
+        Number firstProcess = processor.process(5);
+        resetTimer(processor, 1);
+        Number secondProcess = processor.process(null);
+        resetTimer(processor, 1);
+        Number thirdProcess = processor.process(6);
+        resetTimer(processor, 1);
+        Number fourthProcess = processor.process(7);
 
         assertNull(firstProcess);
         assertNull(secondProcess);
@@ -36,20 +36,20 @@ public class EpochProcessorTest {
 
     @Test
     public void testNullFirstPollCycle() throws InterruptedException {
-        EpochProcessor counter = new EpochProcessor();
-        Number firstProcess = counter.process(null);
-        resetTimer(counter, 1);
-        Number secondProcess = counter.process(5);
-        resetTimer(counter, 1);
-        Number thirdProcess = counter.process(6);
+        EpochProcessor processor = new EpochProcessor();
+        Number firstProcess = processor.process(null);
+        resetTimer(processor, 1);
+        Number secondProcess = processor.process(5);
+        resetTimer(processor, 1);
+        Number thirdProcess = processor.process(6);
 
         assertNull(firstProcess);
         assertNull(secondProcess);
         assertEquals(1.0, thirdProcess.doubleValue(), 0.1);
     }
 
-    private void resetTimer(EpochProcessor counter, int secondsToRemove) {
-        counter.lastTime.setTime(counter.lastTime.getTime() - (secondsToRemove * 1000));
+    private void resetTimer(EpochProcessor processor, int secondsToRemove) {
+        processor.lastTime.setTime(processor.lastTime.getTime() - (secondsToRemove * 1000));
     }
 
 }